Rename folder "parser" to "logical-parser" #51
