+++ internal/storage/files/logical-parser/stack.go
